Move RateLimits method to a service
